+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 Fixed an issue where files with no common content would show duplicate
 hunks.
 
+Fixed a performance issue when files had extremely long lines
+(e.g. 100,000+ characters).
+
 ## 0.59 (released 20th July 2024)
 
 ### Diffing